requirements.txt
altair==5.4.1
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
autopep8==2.3.1
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
executing==2.0.1
fastjsonschema==2.19.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
fonttools==4.48.1
fqdn==1.5.1
h11==0.14.0
htmlmin==0.1.12
httpcore==1.0.7
httpx==0.27.2
idna==3.10
ImageHash==4.3.1
ipykernel==6.29.0
ipython==8.20.0
ipywidgets==8.1.5
iso3166==2.1.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.1.1
json5==0.9.28
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.5
lux-api==0.5.1
lux-widget==0.1.11
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
missingno==0.5.2
mistune==3.0.2
MouseInfo==0.1.3
multimethod==1.12
narwhals==1.14.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.2.2
notebook_shim==0.2.4
numpy==1.26.2
opencv-python==4.8.1.78
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandas-profiling
pandocfilters==1.5.1
parso==0.8.3
phik==0.12.4
pillow==10.2.0
platformdirs==4.1.0
plotly==5.19.0
prometheus_client==0.21.0
prompt-toolkit==3.0.43
psutil==5.9.8
pure-eval==0.2.2
pyarrow==15.0.0
PyAutoGUI==0.9.54
pycodestyle==2.12.1
pycparser==2.22
pydantic<2.0
pydantic_core
PyGetWindow==0.0.9
Pygments==2.17.2
PyJWT==2.9.0
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
pytweening==1.0.7
pytz==2023.4
PyWavelets==1.7.0
pywin32==306
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==25.1.2
referencing==0.33.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scikit-learn==1.3.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==70.3.0
sh==2.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
tangled-up-in-unicode==0.2.0
tenacity==8.2.3
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tornado==6.4
tqdm==4.67.0
traitlets==5.14.1
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==2.2.3
visions==0.7.4
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wheel==0.43.0
widgetsnbextension==4.0.13
